requirements.txt
litellm[proxy]
ollama
pyautogen[retrievechat]
graphrag
chainlit
graphrag
marker-pdf
torch